+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Changelog
 
 ### Bug fixes
 
+* Fixed a crash after shared NSURLSession invalidate
+  [334](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/334)
+
 * Fix visionOS compilation errors. Note that visionOS is not yet officially supported.
   [327](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/327)